+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Use functional components instead of class components ([UIFC-362](https://folio-org.atlassian.net/browse/UIFC-362))
 * Remove dependency to stripes-acq-components, use mocks local ([UIFC-368](https://folio-org.atlassian.net/browse/UIFC-368))
 * Fix GitHub Actions workflow not running for tags ([FOLIO-4086](https://folio-org.atlassian.net/browse/FOLIO-4086))
+* Deleting search string via keyboard does not update search results ([UIFC-358](https://folio-org.atlassian.net/browse/UIFC-358))
 
 ## [6.1.0](https://github.com/folio-org/ui-plugin-find-finc-metadata-collection/tree/v6.1.0) (2024-03-20)
 * Translation updates